Nifti reshaping
Previously, due to the reshaping of the nifti file here, the voxels were left in
a different order than at the beginning, leading to nonsense results after
parcellating the processed cifti. To solve this, here we avoid
some (unnecessary?) reshaping performed in mrutils.import_nifti() and
mrutils.nifti_save(), and introduce also backward reshaping to get back to the
original shape after all processing is done.